IGNITE-19566 Merge ExceptionUtils and IgniteExceptionUtils
