+++ modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/ExchangeServiceImpl.java
@@ -36,7 +36,6 @@ import org.apache.ignite.internal.sql.engine.message.SqlQueryMessagesFactory;
 import org.apache.ignite.internal.sql.engine.util.Commons;
 import org.apache.ignite.internal.util.ExceptionUtils;
 import org.apache.ignite.lang.IgniteException;
-import org.apache.ignite.lang.IgniteExceptionUtils;
 import org.apache.ignite.lang.IgniteInternalCheckedException;
 import org.apache.ignite.lang.IgniteInternalException;
 import org.apache.ignite.sql.SqlException;
@@ -137,7 +136,7 @@ public class ExchangeServiceImpl implements ExchangeService {
         Throwable cause = ExceptionUtils.unwrapCause(t);
 
         if (cause instanceof IgniteException) {
-            return cause == t ? (IgniteException) cause : IgniteExceptionUtils.wrap(t);
+            return cause == t ? (IgniteException) cause : ExceptionUtils.wrap(t);
         } else if (cause instanceof IgniteInternalException) {
             IgniteInternalException iex = (IgniteInternalException) cause;
